// webshop-components/src/components/Table/interfaces.ts
export enum CellStatus {
  Warning = 'Warning',
  Danger = 'Danger',
  Success = 'Success',
  Info = 'Info',
}

export interface IStatusDefinition {
  [statusValue: string]: {
    label: string;
    type?: CellStatus;
    color?: string;
  };
}

export interface IGetRowState<T> {
  (row: IRowInfo<T>): RowState;
}

export interface IColumnAccessorFunction<T, R = string> {
  (data: T): R;
}

export interface IColumnAggregateFunction<V> {
  (values: V[]): V;
}

export interface IFilterFunction {
  (query: { pivotId?: string, id?: string, value: string }, row: any, column: any): boolean;
}

export interface IColumnCellFormatFunction {
  <T>(value: T, row: any): string | JSX.Element; // TODO consult @domagoj.cerjan about types here
}
export interface IColumnCellFormatFunction {
  <T>(value: T, row: any): string | JSX.Element; // TODO consult @domagoj.cerjan about types here
}

export interface ICellEventHandler<T> {
  (event: React.SyntheticEvent<HTMLDivElement | HTMLButtonElement>, row: T, original: IRowInfo<T>): void;
}
// ...
export interface IRowInfo<T> {
  row: any; // #TODO @bigd - type me if possible
  rowValues: any; // #TODO @bigd - type me if possible

  original: T;
  index: number;
  viewIndex: number;
  pageSize: number;
  page: number;
  level: number;
  nestingPath: number[];
  aggregated: boolean;
  groupedByPivot: boolean;
  // subRows: any[]; // - not used
}

interface IInternalRowPluckedValues<A extends any> {
  [key: string]: A;
}
export interface IInternalRowInfo<T> extends IInternalRowPluckedValues<any> {
  _index: number;
  _nestingLevel: number[];
  _original: T;
  // _subRows:  // - not used
  _viewIndex: number;
}

export interface IInlineStyle {
  [key: string]: string | number;
}
export interface ITdProps {
}

export interface IResizedInfo {
}

export interface IActionInfo<T> {
  title?: string | ((original: T) => string);
  tooltip?: string | ((original: T) => string);
  icon?: string;
  className?: string;
  isVisible?: boolean | ((row: any) => boolean);
  disabled?: boolean | ((row: any) => boolean);
  onClick?: ICellEventHandler<T>;
}

export interface IColumnProps<T> {
  actions?: Array<IActionInfo<T>>;
  className?: string;
  styles?: { [key: string]: string | number };
  onClick?: ICellEventHandler<T>;
  onChange?: ICellEventHandler<T>;
  onBlur?: ICellEventHandler<T>;
  url?: string;
  openInNewTab?: boolean;
  constructUrl?: IConstructUrl<T>;
  statusDefinition?: IStatusDefinition;
  urlHandle?: boolean; // added for the link cell
}

export interface ICellInfo<T, V> {
  isExpanded: boolean; // true if this row is expanded
  value: V; // the materialized value of this cell
  resized: IResizedInfo; // the resize information for this cell's column
  show: boolean; // true if the column is visible
  width: number; // the resolved width of this cell
  maxWidth: number; // the resolved maxWidth of this cell
  tdProps: ITdProps; // the resolved tdProps from `getTdProps` for this cell
  columnProps: IColumnProps<T>; // the resolved column props from 'getProps' for this cell's column
  className?: string; // the resolved array of classes for this cell
  style?: IInlineStyle; // the resolved styles for this cell
}

export interface IConstructUrl<T> {
  (original: T, row: IRowInfo<T>): string;
}

export interface ICellInternalProps<T, V> {
  actions?: Array<IActionInfo<T>>;
  accessor?: DeepKeyOf<T> | IColumnAccessorFunction<T, V>;
  aggregate?: IColumnAggregateFunction<V>;
  dateTimeFormat?: string;
  formatter?: IColumnCellFormatFunction;
  field: DeepKeyOf<T> | string;
  label?: string;
  url?: string;
  openInNewTab?: boolean;
  constructUrl?: IConstructUrl<T>;
  onClick?: ICellEventHandler<T>;
  onChange?: ICellEventHandler<T>;
  onBlur?: ICellEventHandler<T>;
  statusDefinition?: IStatusDefinition;
  urlHandle?: boolean; // added for the link cell
}

export interface ICellProps<T, V> extends IRowInfo<T>, ICellInfo<T, V>, ICellInternalProps<T, V> {
}